bugfix
- uint8_t -> uint16_t
- invalid method name
- missing default value (no_profiles = 1)

+++ src/loader/build_footpaths.cc
@@ -439,7 +439,7 @@ void write_footpaths(timetable& tt, uint16_t const& no_profiles) {
   tt.locations_.preprocessing_footpaths_out_.clear();
 }
 
-void build_footpaths(timetable& tt, uint8_t const& no_profiles) {
+void build_footpaths(timetable& tt, uint16_t const& no_profiles) {
   add_links_to_and_between_children(tt);
   link_nearby_stations(tt);
   transitivize_footpaths(tt);
